Organize common mqtt constants
